+++ rollup.config.js
@@ -1,5 +1,25 @@
 /*
- * Copyright (c) 2018 Rafael da Silva Rocha.
+ * Copyright (c) 2018-2019 Rafael da Silva Rocha.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
  */
 
 /**
@@ -7,31 +27,25 @@
  * @see https://github.com/rochars/ieee754-buffer
  */
 
-import babel from 'rollup-plugin-babel';
-import fs from 'fs';
-
-const head = ";(function (global, factory) {" +
-  "typeof exports === 'object' && typeof module !== 'undefined' ? module.exports = factory() :" +
-  "typeof define === 'function' && define.amd ? define(factory) :" +
-  "(global.IEEE754Buffer = factory());" +
-"}(this, (function () {";
-const foot = "return IEEE754Buffer;})));"
+import compiler from '@ampproject/rollup-plugin-closure-compiler';
 
 export default [
   {
-    input: 'ieee754-buffer.js',
+    input: 'index.js',
     output: [
       {
-        file: 'ieee754-buffer.umd.js',
+        file: 'dist/ieee754-buffer.js',
         name: 'IEEE754Buffer',
-        format: 'iife',
-        strict: false,
-        banner: head,
-        footer: foot
-      }
+        format: 'umd'
+      },
     ],
     plugins: [
-      babel()
+      compiler({
+        language_in: 'ECMASCRIPT6',
+        language_out: 'ECMASCRIPT3',
+        compilation_level: 'SIMPLE',
+        warning_level: 'VERBOSE'
+      })
     ]
-  },
+  }
 ];