+++ test/loader.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2018 Rafael da Silva Rocha.
+ * Copyright (c) 2018-2019 Rafael da Silva Rocha.
  * https://github.com/rochars/ieee754-buffer
  *
  */
@@ -9,14 +9,14 @@ let ieee754Buffer;
 // UMD bundle
 if (process.argv[3] == '--umd') {
 	console.log('umd tests');
-	ieee754Buffer = require('../ieee754-buffer.umd.js');
+	ieee754Buffer = require('../dist/ieee754-buffer.js');
 
 // Source
 } else {
 	require = require("esm")(module);
 	global.module = module;
 	console.log('Source tests');
-	ieee754Buffer = require('../ieee754-buffer.js');
+	ieee754Buffer = require('../index.js');
 }
 
 module.exports = ieee754Buffer;